backend/requirements.txt
fastapi[standard]
uvicorn
sqlalchemy
pydantic[email]
pydantic-settings
openai
python-dotenv
supabase
celery
requests
redis
ruff
pymysql
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pyjwt
langchain
langchain-core
langchain-text-splitters
limits

loguru
oss2

langsmith==0.3.13
pydantic==2.10.6
pydantic_core==2.27.2
pydantic-settings==2.6.1
